Implement additional SRMD support
The support for SRMD is basically complete, except for lack of support
for Annotations. This is a wider problem, since most SSP elements can
contain Annotations.

There is (experimental) support for custom classification parsers. This
is intended to support usecases like MIC Core, where there is a defined
set of keywords, with values that may require additional parsing, such
as dates to be useful.

Currently, SRMD __init__ methods are implemented in a slightly weird
way, where they either take an XML element or a str as their first
argument. This is for consistency with other classes in the project.