Add ability to create empty SSP, FMU archives
Add new 'mode' parameter to ZIPFile, with the plan to eventually
deprecate the readonly parameter. For now, that parameter works as
previously, but emits a DeprecationWarning when used, with a suggested
code change to use the mode parameter instead.

Using mode "w", or "a" if the target file doesn't exist loads an empty
zip file. This will later be used to implement functionality to create
SSP archives from scratch.